Task_Management/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
jwcrypto==1.5.6
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1